+++ Source/Urho3D/IO/VectorBuffer.h
@@ -22,9 +22,9 @@ public:
     VectorBuffer(Deserializer& source, i32 size);
 
     /// Read bytes from the buffer. Return number of bytes actually read.
-    unsigned Read(void* dest, unsigned size) override;
+    i32 Read(void* dest, i32 size) override;
     /// Set position from the beginning of the buffer. Return actual new position.
-    unsigned Seek(unsigned position) override;
+    i64 Seek(i64 position) override;
     /// Write bytes to the buffer. Return number of bytes actually written.
     i32 Write(const void* data, i32 size) override;
